Tercer commit

+++ app/src/main/java/com/example/gomezgomezjosemaria/model/Jugador.kt
@@ -3,7 +3,7 @@ package com.example.gomezgomezjosemaria.model
 data class jugador(
     val id: String = "",
     val nombre: String = "",
-    val numero: Double = 0.0,
+    val numero: Int = 0,
     val posicion: String,
     val descripcion: String = ""
 )
